requirements.txt
django>=5.0.1,<6.0.0 # Dja;ngo Framework
djangorestframework>=3.14.0,<4.0.0  # DRF (Django Rest Framework)
psycopg2-binary>=2.9.9,<3.0